--- constants.go
@@ -1,22 +1,0 @@
-package main
-
-const (
-	port                    = ":8888"
-	databaseString          = "ConduitDB"
-	connectionString        = "mongodb://localhost/"
-	userTableInfoString     = "userInfo"
-	userProjectInfoString   = "userProj_"
-	joinedProjectInfoString = "joinedProj_"
-	coJoinProjectInfoString = "coJoinProj_"
-
-	checkUID        = "uid"
-	checkGender     = "gender"
-	checkPID        = "pid"
-	checkJUID       = "juid"
-	checkCUID       = "cuid"
-	checkProfilePic = "profile_pic"
-	checkName       = "name"
-
-	registerOperation      = "REGISTER"
-	updateProfileOperation = "UPDATE_PROFILE"
-)
